tiltfile: move some more converters into tiltfile/value
