+++ internal/tiltfile/io/blob.go
@@ -4,6 +4,8 @@ import (
 	"fmt"
 
 	"go.starlark.net/starlark"
+
+	"github.com/windmilleng/tilt/internal/tiltfile/value"
 )
 
 type Blob struct {
@@ -15,6 +17,10 @@ var _ starlark.Value = Blob{}
 
 func NewBlob(text string, source string) Blob {
 	return Blob{Text: text, Source: source}
+}
+
+func (b Blob) ImplicitString() string {
+	return b.Text
 }
 
 func (b Blob) String() string {
@@ -34,3 +40,5 @@ func (b Blob) Truth() starlark.Bool {
 func (b Blob) Hash() (uint32, error) {
 	return 0, fmt.Errorf("unhashable type: blob")
 }
+
+var _ value.ImplicitStringer = Blob{}
